Monad Transformers to the rescue!
**Why all this?**

Honestly, I don't quite remember. It started out as the question of how
to modularise out networking, which lead to the realisation that
"consensus" (the `Protocol` class) would handle too many things at the
same time, which led to a major attempt to draw some interface
boundaries.

**What's all this?**

We ended up re-formulating consensus, storage (mainly: mempool and
blockstore), and networking as a monad transformer stack, using a
bunch of "mtl-style" effect classes to convey which layer owns resp. has
access to which concerns. The main classes are `MonadProtocol`,
`MonadMempool`, `MonadBlockStore`, and `MonadNetwork`. There are
also some classes around read/write access to the "World State"
(aka "state tree"), but that hasn't been fleshed out fully yet.

We have a concept of a `Node` monad, which provides mempool and
blockstore access (as well as state tree). The impure (IO)
implementation is `Oscoin.Node.NodeT`, for testing we have
`Oscoin.Consensus.Test.Node.TestNodeT` (which is just a `StateT`).
Consensus (namely `NakamotoT` and `SimpleT`) implement `MonadProtocol`
with constraints such that they can access the mempool and blockstore.

**Are we done yet?**

Hah, no! Some things off the top of my head:

* things could be named a bit better (eg. `addTxs` vs. perhaps `mempoolAddTxs`?)
* we don't know what a transaction is, and I understand that it will be
  radicle's job to interpret whatever bytes come in into whatever a
  transaction means. We do, however, currently let network parse the bytes
  into some transaction type, which requires to carry a phantom type for
  both `NetworkT` and `NodeT`. While this allows to pick the type when
  running the stack, which is handy for our "dummy" executables, it is
  also unsafe: you can run the same stack at a different type. I think a
  transaction should just be a bytestring, and whatever the interface to
  radicle will be would just carry an interpretation function.
* strikingly, the implementation for the `step` function (now `stepM`)
  is exactly the same for both `NakamotoT` and `SimpleT`. This seems
  like a hint that we haven't yet nailed the interface for "consensus" -
  perhaps the state manipulation (mempool, blockstore) on incoming
  messages should be handled generically by `NodeT`, and consensus is
  just `tick`? This would also solve the concurrency story: mining
  should be running in a separate thread, and the required state for
  doing this wouldn't need to be shared with the recv/send loop.
* We should probably have tests which ensure that the pure `TestNodeT`
  and the `IO` `NodeT` behave the same according to their `MonadXXX`, as
  soon as we implement some actual persistent storage. Or rethink how the
  simulator tests work ("state machine testing"?)
* I'm not convinced (anymore) that a separate network (transformer)
  layer is needed, but let's see until after something more
  production-like is there (otherwise, the simulator should probably
  exercise it, too)