Add 'Oscoin.API.Client' type class
Add a type class that defines the abstract interface of a client for a
Oscoin node. We will follow this up with a concrete implementation of a
client for the HTTP API.

This commit also moves all types used by the client into the
`Oscoin.API.Types` module. These types are not specific to the
implementation (i.e. HTTP). As part of this we removed the `ok` and
`err` constructor aliases for API results. We now simply use the `Ok`
and `Err` constructors.