Use upstream version of 'wai-extra' instead of fork
Our forked version of `wai-extra` only added [this commit][1] which
added the `HasCallstack` constraint to assertions. The intention is
that, if an assertion fails, the test runner prints the location in the
test instead of the location of the assertion code.

However we currently do not use any of the assertion functions
exposed by `Network.Wai.Test`. We use wrapped functions from
`Oscoin.Test.HTTP.Helpers` that replicate the same behavior. This means
there is no need to use the forked version of `wai-extra`.

[1]: https://github.com/cloudhead/wai/commit/a3464cc76198d51ae7cf838690eadeeefd627f9c