CLI uses keypair from file
With this commit the CLI commands will be able to read the users keypair
from disk. We also introduce a `generate-keypair` command that creates a
new keypair and writes it to disk.

To implement this we chose to add another `MonadKeyStore` class and add
it as a constraint to `runCommand`. This approach was chosen to
facilitate testing: Almost all of the tests we will write for the CLI
will not be concerned with key pairs. We just want to ensure that a key
pair is avialable. To this end we provide a lightweight instance for
`MonadKeyStore` that just reads a key pair from memory.