Remove unused files from cache
We add a CI command that deletes unused large files from the folder that
is used as a build cache.

We also update the cache version to rebuild it from scratch and not
include unused data from previous caches.

All this results in a compressed cache size of 180MB instead of 1.2GB
and improves CI times significantly.