Delete failed TXs when mining block
When a transaction fails we want to delete it from the Mempool. This
requires a refactoring of the consensus code.

We simplify the miner function to `StdGen -> BlockHeader a -> Maybe
(BlockHeader a)`. The code that was previously `mineBlock ::
NakamotoMiner tx s` is moved into `stepM` and the new miner function is
factored out.

We also change the `mineBlockRandom` used for tests to satisfy the new
interface and move the code to the test.

As a follow-up we will log failed transactions and store the result of
failed transactions somewhere.