Resolve radicle references when querying state
When we query the blockchain state through the API we want to resolve
references. This means if the queried value binds to a reference we
get the value the Reference points to instead of the reference itself.

We also do some cleanup of the test helper functions to avoid code
duplication.