Eliminate use of 'Default' in 'Block' functions
We change the implementation of `genesisBlock` and `emptyGenesisBlock`
to not make use of the `Default s` constraint for the state type
parameter `s`. Instead we require an explicit state argument.

For the tests we replace the `Default s` constraint with an `Arbitrary
s` constraint. The only exception is `arbitraryValidBlockchain` where
the state argument is explicit. This is required because the tests fail
otherwise.