Turn NakamotoT from RWST to StateT
We change the definition of `NakamotT` from
```hs
newtype NakamotoT tx s m a = NakaomotoT (RWST (NakamotoEnv tx s) () StdGen m a)
```
to
```hs
newtype NakamotoT tx s m a = NakaomotoT (StateT StdGen m a)
```

`NakamotEnv` is eliminated because we always run `NakamotoT` with a
fixed environment. This is now hardcoded as `nakConsensus`.