Align 'stepM' code by eliminating msg complexity test
The goal of this commit is to make the `stepM` method the same for both
Nakamot and Simple consensus in the tests. There was a subtle difference
in the previous implementation where the Simple consensus would not
propagate messages. Changing this breaks the message complexity tests
which we also remove since they provide no value.

This all allows us to unify the `stepM` code in `applyMessage` and
eliminate `stepM` as a method. Note that we always use `identityEval` in
`applyMessage`. Theoretically the Nakamot allows configuration of the
evaluator, but this was not used and will be removed in the future.