Eliminate 'evals' and 'applyValidExprs'
We eliminate `evals` and `applyValidExprs` from
`Oscoin.Consensus.Evaluator` and `genesisBlock` from
`Oscoin.Crypto.Blockchain.Block`. `genesisBlock` was only a transitive
dependency of `testOscoinBlockchain`. We eliminated that test since it
only tested the block and blockchain validation, both of which are not
properly implemented yet. We will add these tests again when we have a
proper implementation.