CI: Publish Docker images
This commit makes our CI pipeline publish Docker images for each of the
executables we build.

When not running on macOS, the executables are statically built to
avoid the runtime dependencies on shared libraries and thus making it
easier to create a functioning and slim runtime Docker image.

Since the final binaries are to be released in Docker images, we now do
optimized compilation. While this increases the compilation time, it
reduces the run-time of the tests.

Fixes #248