P2P: Noise Handshake
Implements [Noise Protocol](https://noiseprotocol.org/noise.html) handshakes,
which allows gossip to be encrypted and perform authentication (that is, verify
that a peer presenting a public key indeed owns the corresponding private key).

It is possible to use this for HTTP communication as well, but this is not
implemented in this PR.

**Why do we need this?**

Gossip protocols are somewhat easier to attack than DHT networks, as
RPCs can influence the topology more directly. Also, DHTs (that is,
Kademlia variants) implement a reputation score, which makes it harder
to separate trustworthy nodes from the network (although the metric for
trustworthiness is merely uptime). Given authenticated nodes, we can,
for example:

* Implement a reputation score based on arbitrary metrics (eg. stake)
* Run a closed network.
* Run semi-closed (sub-)networks, where peers can verify each other's identity.
* Run consensus protocols based on notarisation or BFT.
* Impement off-the-record messaging.

**Notes**

* The `NodeId` now carries the full `PublicKey` (vs. only its hash), which is
  needed for signing / verifying protocol messages.
* The default `secureHandshake` currently runs the Noise `NN` handshake pattern,
  which does **not** perform authentication. Instead, we use it to establish a
  secure channel, over which we exchange our usual ECDSA public keys, which are
  subsequently used for [Channel Binding](https://noiseprotocol.org/noise.html#channel-binding).
  This was chosen in order to simplify things for starters, as we would
  otherwise need to maintain an additional DH keypair per node.

    A future implementation, however, should consider implementing some
    variant of ["Noise Pipes"](https://noiseprotocol.org/noise.html#noise-pipes),
    potentially employing [dummy keys](https://noiseprotocol.org/noise.html#dummy-keys),
    because:

    * it allows for 0-RTT handshakes
    * it allows to conceal which handshake pattern was used
    * and thus, offers more flexibility to evolve the protocol in the future

Fixes #215