CI: Sanitize Docker tags with slashes
With #260, we introduce publishing of Docker images in CI, but,
soon after merging it in, @geigerzaehler found out that branch names
with forward slashed were failing to build due to Docker tags not
supporting them.

This PR makes the branch name Docker tag be computed with slashed
replaced with dashes.