Migrate to oscoin/gossip for the gossip protocol.
The code implementing epidemic broadcast trees and the HyParView
protocol were being shipped as part of this repo, but since then the
code was migrated into a repository of its own (oscoin/gossip).

This commit ports over the code to use the external library, removing
all the references to the previous version in use.