Allow the keys directory to be customisable.
This (fairly invasive) commit changes the MonadKeyStore and most of the
CLI parsers to take an extra "--keys" option to pass a user-specified
directory where to store the oscoin keys.

This is needed so that we can run the integration tests on CI without
messing around with any key the user might already have on its home
directory.