Add more events and report them.
This commit (and PR) builds on top of #342.

This commit adds more events to our NotableEvent enumeration, which is
now a GADT so that we can express existentials, which are useful
sometimes, especially when dealing with polymorphic things like
transactions.

On top of this, now 'applyBlock' and 'applyTx' yield a richer type,
namely a '([NotableEvent], a)', so that upstream callers can correctly
emit the events. This also relax the need for the specific monad 'm' to
implement 'MonadIO', which would have destroyed any hope of testing the
code in a completely pure setting.

Last but not least, I did catch a bug in 'newMetricsStore' where I was
not correctly tying the recursive knot for the RTS gauges, meaning that
data was not getting correctly updated.