(Allegedly) improve the logger
This commit tries to bring our Oscoin.Logging.Logger closer to the
logfmt format and it also tries to adhere to the recommended best
practices outlines at https://brandur.org/logfmt.

Last but not least, it introduces the concept of a `StyleFormatter`
(which can be derived, amongst the other things, from an `Environment`)
to style the logging to be more or less human friendly. The term
`StyleFormatter` has been borrowed by [logrus]().

For an extra touch of hipsterism, (optional) coloured output is
supported, too.