Fix #373 (remove hardcoded size in generators)
This commit reshuffles a bit the generators and arbitrary instances we
had for blocks and blockchains, as well as fixing #373 properly (i.e. do
not hardcoded the size of the chain in the generator, but rather use
'sized' and grab the seed out of the continuation).

I went back and forth trying to find a satisfactory "split" for the
generators I wrote as part of #320 and what we had already, mostly
because for some data structures it doesn't make _too much_ sense to
generate an Arbitrary value in vacuum, as usually blocks depends from
their predecessors. Finally, I did settle on this logical split, which I
am open to discuss:

- Foo.Bar.Arbitrary contains either proper `Arbitrary` instances OR
  generators which do not guarantee that the stuff we generate is
  semantically valid. For example, calling 'arbitraryBlock' clearly
  generates a well-formed block which is not linked to any parent (or
  any chain) and therefore is not fully valid, in this sense.

- Foo.Bar.Generators contains proper generators, i.e. functions in the
  'Gen' monad which not only do generate syntactically-valid structures,
  but they also go the extra mile by providing interesting samples. An
  example falling into this bucket would be the 'genOrphanChains'
  generator.

Let me know if this is a bit of a stretch or it might lead to confusion
down the road!