Dump xxhash, hello fnv1a
This commit dump xxhash in favour of fnv1-a. The reason was incidental
as I was trying to see if the latter yielded less collisions
(disclosure: it doesn't) but as I could piggyback on 'memory' dropping
one dependency is always good. Plus, the running times doesn't seem to
be so drastically different.

Finally, the Arbitrary instance for Rad.Value has been restored to the
simple one, and the instance for 'Tx' has been made more complicated.