tests: Only test CBOR HTTP content type
We only test the CBOR encoding in the HTTP tests and drop all other
combinations. This reduces the test time dramatically.

Testing more combinations than CBOR/CBOR only covers more of the
encoding/decoding and content type negotiation code. For the near future
we will only use the CBOR content type and so this loss of coverage is
acceptable.