Replace RadTx with DummyTx in BlockStore tests
We replace the use of `RadTx` with `DummyTx` in the block store tests.
This is one step towards getting rid of Radicle. The change also
improves the run time of the block store tests by roughly 40%.

To achieve the transition we start by extracting `DummyTx` into the
`Test.Oscoin.DummyLedger` module.

The interface for the SQLite storage had a `SQLite.ToRow tx` constraint
for all transactions that was implemeted for `Tx c Radicle.Value`. A
naive implementation of `SQLite.ToRow DummyTx` would lead to
incompatibilities with the database schema. For this reason I introduced
the `IsTxRow` class. This constraint needs to be satisfied in
order to store transactions in the database.