Introduce 'Ledger' for Node, Miner, and tests
We introduce the `Ledger` abstraction which wraps the block store, state
store, and receipt store. The purpose of the `Ledger` is to provide
consistent access to block artifacts (states and receipts). This commit
only integrates the ledger with `Oscoin.Node`,
`Oscoin.Consensus.Mining`, and the `Oscoin.Test.Consensus.Nakamoto`. It
remains to integrate the ledger with `Protocol` where it will replace
the `BlockStore`.