Use dictionary-style API for ReceiptStore
Continuing the work of #453 we implement a dictionary style API intended
to replace `MonadReceiptStore`. We do not yet eliminate the
`MonadReceiptStore` class. This is left for a future PR.

To accomplish the change we introduce `ContentStore`, a generalized
version of `HashStore`. `ContentStore` is an interface for a mutable
key-value store where the key is computed from the value.