Use dumb in-memory Blockstore in API tests
Before, the Node API tests used the STM blockstore that does chain
selection *and* the protocol that is now responsible for chain
selection. We change the block store for the API tests to a dumb
in-memory blocks store. This is a step towards resolving #460.

Also removes redundant `BlockStore.Pure` instances