Callstacks for debug and development prelude functions
We update `Oscoin.Prelude` to allow for a better development experience
when using `notImplemented` and `undefined`.

During development `notImplemented` and `undefined` are often used to
leave holes in the code. When running tests against code that contains
multiple `notImplemented` statements we want to know which of these
threw an error. For this we add the `HasCallstack` constraint to
`notImplemented` and `undefined`.

We also start exporting `Protolude.error` from `Oscoin.Prelude` so that
we can put some more information into errors during development.