Cleanup 'Oscoin.Crypto.Blockchain.Block'
This commit includes changes that improve the code in the
`Oscoin.Crypto.Blockchain.Block` module.

We eliminate `genesisBlock`. This function was only used in one place
where it was easily replaced by `emptyGenesisFromState`.

We eliminate the unused `withHeader` function.

We rewrite `linkParent` and `sealBlock` using lenses. This simplifies
the code an reduces duplication. But more importantly we now guarantee
the correctness of the `blockHash` field by using `blockHeaderL` instead
of relying on us setting the field manually.