Add beneficiary to blocks
Add the *beneficiary* to the block. That is, the account that should be
receiving block rewards.

This just adds the ability to have a block beneficiary, but does not
dispense any rewards. That will be a separate PR.

A new flag is introduced: `--beneficiary`, which takes an `Address` that
is parsed and converted into a `PublicKey` and stored in the node
`Config`. It is then passed to the miner for inclusion in blocks.

A new data type is introduced: `BlockData`, which encapsulates
beneficiary and transactions.  The beneficiary (a `PublicKey`) is hashed
alongside the transactions, for tamper-evidence. This means that now,
`blockDataHash = merklehash(beneficiary || txs)`.