Make API polymorphic in 'tx'
The code in `Oscoin.API.HTTP.Handlers` and related modules is now
polymorphic in the transaction type. See #552.

As part of this change we also eliminate the `getAllTransactions`
endpoint. It’s implementation did not use a serialisation of a
transaction list but rather a serialisation of the Mempool instance.
This coupled the API to the concrete mempool instance used by the node.