Move transaction validation to Mempool
Before, transaction validation only happened in the API Handlers when a
transaction was submitted. We now validate a transaction when it is
inserted into the Mempool.

We replace the `Mempool.addTxs` with `Mempool.addTx`. Inserting multiple
txs at once was never properly used.

Now, the API does not know how a transaction is validated. It only gets
the result of that validation.

We also start rejecting invalid transactions from gossip.

Fixes #553