Extend API.Client test, remove redundant tests
We add a test for invalid transactions with
`Oscoin.API.Client.submitTransaction` in `Test.Oscoin.API`. This allows
us to remove the low-level test from `Test.Oscoin.API.HTTP`.

We also remove the so called “smoke tests” from the API. This behavior
is also covered by the tests for `Oscoin.API.Client.submitTransaction`.