cli: Prepare genesis creation for OscoinTx
To run the node with OscoinTx (#568) we need to create a genesis block
for the Oscoin ledger. We prepare for this by removing the inclusion of
transactions in the genesis block. While we might want to have that
feature later the current implementation will not work with the Oscoin
ledger. The reason is that currently transactions are signed with the
node key, but we need to sign it with a fixed account owner key.