Generate genesis block from parameters (#613)
Before this commit the node loaded the genesis block from a YAML file.
This required us to be able to serialise a whole block as YAML. Now the
node only loads the `GenesisParameters` from a YAML file and uses
the parameters to build the genesis block.

Conversely, we change the `genesis` command of the CLI to create these
`GenesisParameters` from command line arguments.

As a consequence we can remove a multitude of `ToJSON` and `FromJSON`
instances and close #602. Note that there still remain some instances
for example for `Crypto.ShortHash` that are not trivial to get rid of.

The change also makes it possible to build the genesis block for every
type of transaction (since the transaction set is empty). This closes #589.