Reusable keypair file storage
We extract the functionality to read and write key paris to disk from
`Oscoin.CLI.KeyStore` into `Oscoin.Crypto.PubKey.FileStore`. This makes
this functionality easier to use since we are providing just functions
instead of an MTL-style type class.

This is a necessary step towards having separate key management for the
node keys and the account keys (#578).