Remove 'getTransaction' API
We remove `getTransaction` from the node’s HTTP API.

The endpoint in its current is not needed for anything. It also does not
match the specification of that endpoint documented in the specs.
Instead of carrying this dead weight around we are removing it.

This was briefly discussed with and okayed by @cloudhead.