Short hash parsing and serialising
The account CLI needs to display and parse human readable short hashes
(account IDs). To make this easy I propose a couple of changes.

Currently, the `HasHashing` class provides a method to parse a short
hash from `Text`. A partial inverse of this function only exists as
`formatShortHash` which returns a formatter. It is not clear that these
functions are inverse to each other and this is nowhere enforced.

I propose to decouple the byte representation of a short hash from the
human readable representation (encoding) of those bytes. In particular
`parseShortHash` is removed from the `HasHashing` type class and
replaced with a `Binary (ShortHash c)` constraint.

This has the additional advantage that the encoding is independent of
the type class. We also make it easier to implement different
polymorphic encoding/decoding functions for short hashes in the future

The same approach could be adapted for `Hash`, too.